requirements.txt
pytest
python-dotenv==1.0.1
Flask==3.0.3
flask-wtf==1.2.1
password-validator==1.0
better-profanity==0.7.0
SQLAlchemy==2.0.33
pathlib2~=2.3.7.post1